+++ examples/nas/oneshot/spos/search.py
@@ -1,17 +1,88 @@
-# This file is to demo the usage of multi-trial NAS in the usage of SPOS search space.
-
-import click
+import time
 import json
-import nni.retiarii.evaluator.pytorch as pl
-import nni.retiarii.strategy as strategy
-from nni.retiarii import serialize
-from nni.retiarii.experiment.pytorch import RetiariiExeConfig, RetiariiExperiment
-from torchvision import transforms
-from torchvision.datasets import CIFAR10
-from nn_meter import load_latency_predictor
+import random
+import logging
+import argparse
+import numpy as np
+import torch
+import torchvision.transforms as transforms
+import torchvision.datasets as datasets
 
-from network import ShuffleNetV2OneShot
-from utils import get_archchoice_by_model
+import nni
+from nn_meter import load_latency_predictor
+import nni.retiarii.nn.pytorch as nn
+import nni.retiarii.strategy as strategy
+from nni.retiarii.evaluator.functional import FunctionalEvaluator
+from nni.retiarii.utils import original_state_dict_hooks
+from nni.retiarii.oneshot.pytorch.utils import AverageMeterGroup
+from nni.retiarii.experiment.pytorch import RetiariiExeConfig, RetiariiExperiment
+
+from network import ShuffleNetV2OneShot, load_and_parse_state_dict
+from utils import CrossEntropyLabelSmooth, accuracy, ToBGRTensor, get_archchoice_by_model
+
+logger = logging.getLogger("nni.spos.search")
+
+
+def retrain_bn(model, criterion, max_iters, log_freq, loader):
+    with torch.no_grad():
+        logger.info("Clear BN statistics...")
+        for m in model.modules():
+            if isinstance(m, nn.BatchNorm2d):
+                m.running_mean = torch.zeros_like(m.running_mean)
+                m.running_var = torch.ones_like(m.running_var)
+
+        logger.info("Train BN with training set (BN sanitize)...")
+        model.train()
+        meters = AverageMeterGroup()
+        for step in range(max_iters):
+            inputs, targets = next(iter(loader))
+            inputs, targets = inputs.to('cuda'), targets.to('cuda')
+            logits = model(inputs)
+            loss = criterion(logits, targets)
+            metrics = accuracy(logits, targets)
+            metrics["loss"] = loss.item()
+            meters.update(metrics)
+            if step % log_freq == 0 or step + 1 == max_iters:
+                logger.info("Train Step [%d/%d] %s", step + 1, max_iters, meters)
+
+
+def test_acc(model, criterion, log_freq, loader):
+    logger.info("Start testing...")
+    model.eval()
+    meters = AverageMeterGroup()
+    start_time = time.time()
+    with torch.no_grad():
+        for step, (inputs, targets) in enumerate(loader):
+            inputs, targets = inputs.to('cuda'), targets.to('cuda')
+            logits = model(inputs)
+            loss = criterion(logits, targets)
+            metrics = accuracy(logits, targets)
+            metrics["loss"] = loss.item()
+            meters.update(metrics)
+            if step % log_freq == 0 or step + 1 == len(loader):
+                logger.info("Valid Step [%d/%d] time %.3fs acc1 %.4f acc5 %.4f loss %.4f",
+                            step + 1, len(loader), time.time() - start_time,
+                            meters.acc1.avg, meters.acc5.avg, meters.loss.avg)
+    return meters.acc1.avg
+
+
+def evaluate_acc(class_cls, criterion, args, train_dataset, val_dataset):
+    model = class_cls()
+    with original_state_dict_hooks(model):
+        model.load_state_dict(load_and_parse_state_dict(), strict=False)
+    model.cuda()
+    
+    train_loader = torch.utils.data.DataLoader(train_dataset, batch_size=args.train_batch_size, num_workers=args.workers)
+    test_loader = torch.utils.data.DataLoader(val_dataset, batch_size=args.test_batch_size, num_workers=args.workers)
+
+    acc_before = test_acc(model, criterion, args.log_frequency, test_loader)
+    nni.report_intermediate_result(acc_before)
+
+    retrain_bn(model, criterion, args.train_iters, args.log_frequency, train_loader)
+    acc = test_acc(model, criterion, args.log_frequency, test_loader)
+    assert isinstance(acc, float)
+    nni.report_intermediate_result(acc)
+    nni.report_final_result(acc)
 
 
 class LatencyFilter:
@@ -37,41 +108,64 @@ class LatencyFilter:
         return latency < self.threshold
 
 
-@click.command()
-@click.option('--port', default=8081, help='On which port the experiment is run.')
-def _main(port):
-    base_model = ShuffleNetV2OneShot(32)
-    base_predictor = 'cortexA76cpu_tflite21'
-    transf = [
-        transforms.RandomCrop(32, padding=4),
-        transforms.RandomHorizontalFlip()
-    ]
-    normalize = [
-        transforms.ToTensor(),
-        transforms.Normalize([0.49139968, 0.48215827, 0.44653124], [0.24703233, 0.24348505, 0.26158768])
-    ]
-    # FIXME
-    # CIFAR10 is used here temporarily.
-    # Actually we should load weight from supernet and evaluate on imagenet.
-    train_dataset = serialize(CIFAR10, 'data', train=True, download=True, transform=transforms.Compose(transf + normalize))
-    test_dataset = serialize(CIFAR10, 'data', train=False, transform=transforms.Compose(normalize))
+def _main():
+    parser = argparse.ArgumentParser("SPOS Evolutional Search")
+    parser.add_argument("--imagenet-dir", type=str, default="./data/imagenet")
+    parser.add_argument("--checkpoint", type=str, default="./data/checkpoint-150000.pth.tar")
+    parser.add_argument("--spos-preprocessing", action="store_true", default=False,
+                        help="When true, image values will range from 0 to 255 and use BGR "
+                             "(as in original repo).")
+    parser.add_argument("--seed", type=int, default=42)
+    parser.add_argument("--workers", type=int, default=6)
+    parser.add_argument("--train-batch-size", type=int, default=128)
+    parser.add_argument("--train-iters", type=int, default=200)
+    parser.add_argument("--test-batch-size", type=int, default=512)
+    parser.add_argument("--log-frequency", type=int, default=10)
+    parser.add_argument("--label-smoothing", type=float, default=0.1)
 
-    trainer = pl.Classification(train_dataloader=pl.DataLoader(train_dataset, batch_size=64),
-                                val_dataloaders=pl.DataLoader(test_dataset, batch_size=64),
-                                max_epochs=2, gpus=1)
+    args = parser.parse_args()
 
-    simple_strategy = strategy.RegularizedEvolution(model_filter=LatencyFilter(threshold=100, predictor=base_predictor),
-                                                    sample_size=1, population_size=2, cycles=2)
-    exp = RetiariiExperiment(base_model, trainer, strategy=simple_strategy)
+    # use a fixed set of image will improve the performance
+    torch.manual_seed(args.seed)
+    torch.cuda.manual_seed_all(args.seed)
+    np.random.seed(args.seed)
+    random.seed(args.seed)
+    torch.backends.cudnn.deterministic = True
+
+    assert torch.cuda.is_available()
+
+    base_model = ShuffleNetV2OneShot()
+    base_latency_predictor = 'cortexA76cpu_tflite21'
+    criterion = CrossEntropyLabelSmooth(1000, args.label_smoothing)
+    if args.spos_preprocessing:
+        trans = nni.trace(transforms.Compose)([
+            nni.trace(transforms.RandomResizedCrop)(224),
+            nni.trace(transforms.ColorJitter)(brightness=0.4, contrast=0.4, saturation=0.4),
+            nni.trace(transforms.RandomHorizontalFlip)(0.5),
+            nni.trace(ToBGRTensor)(),
+        ])
+    else:
+        trans = nni.trace(transforms.Compose)([
+            nni.trace(transforms.RandomResizedCrop)(224),
+            nni.trace(transforms.ToTensor)()
+        ])
+    train_dataset = nni.trace(datasets.ImageNet)(args.imagenet_dir, split='train', transform=trans)
+    val_dataset = nni.trace(datasets.ImageNet)(args.imagenet_dir, split='val', transform=trans)
+
+    evaluator = FunctionalEvaluator(evaluate_acc, criterion=criterion, args=args, train_dataset=train_dataset, val_dataset=val_dataset)
+    evolution_strategy = strategy.RegularizedEvolution(
+        # model_filter=LatencyFilter(threshold=100, predictor=base_latency_predictor),
+        sample_size=1, population_size=2, cycles=2)
+    exp = RetiariiExperiment(base_model, evaluator, strategy=evolution_strategy)
 
     exp_config = RetiariiExeConfig('local')
     exp_config.trial_concurrency = 2
     exp_config.trial_gpu_number = 1
     exp_config.training_service.use_active_gpu = False
     exp_config.execution_engine = 'base'
-    exp_config.dummy_input = [1, 3, 32, 32]
+    # exp_config.dummy_input = [1, 3, 224, 224]
 
-    exp.run(exp_config, port)
+    exp.run(exp_config, 8084)
 
     print('Exported models:')
     for i, model in enumerate(exp.export_top_models(formatter='dict')):
@@ -79,6 +173,5 @@ def _main(port):
         with open(f'architecture_final_{i}.json', 'w') as f: 
             json.dump(get_archchoice_by_model(model), f, indent=4)
 
-
-if __name__ == '__main__':
+if __name__ == "__main__":
     _main()
